+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 - Import `esbuild` dependency from `npm` to avoid permissions issues [#795].
 - Updated dependencies: `deno-loader`, `sass`, `xml`, `unocss` and icons.
 - Vento components shouldn't inherit the page's data [#797].
+- `icons` plugin: ensure the SVG code includes the `viewBox` attribute.
 
 ## [3.1.2] - 2025-11-18
 ### Added